init: move to new package: cmd/wksctl/init
